pyproject.toml
[project]
name = "ai-trading-team"
version = "0.1.0"
description = "Multi-agent cryptocurrency trading bot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "binance-sdk-derivatives-trading-usds-futures>=5.0.0",
    "langchain>=1.2.3",
    "langchain-anthropic>=1.3.1",
    "python-dotenv>=1.2.1",
    "talipp>=2.7.0",
    "textual>=7.1.0",
    "textual-dev>=1.8.0",
    "weex-sdk>=1.0.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_trading_team"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["ai_trading_team"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
mypy_path = "src"
packages = ["ai_trading_team"]
